pl_develop & split pipes
