+++ requirements/devel.txt
@@ -6,5 +6,3 @@
 
 # extended list of dependencies dor development and run lint and tests
 -r ./test.txt
-
-cloudpickle>=1.2
