+++ tests/models/data/horovod/train_default_model.py
@@ -23,6 +23,7 @@ import sys
 
 import horovod.torch as hvd
 
+
 PATH_HERE = os.path.abspath(os.path.dirname(__file__))
 PATH_ROOT = os.path.join(PATH_HERE, '..', '..', '..', '..')
 sys.path.insert(0, os.path.abspath(PATH_ROOT))
@@ -30,7 +31,8 @@ sys.path.insert(0, os.path.abspath(PATH_ROOT))
 from pytorch_lightning import Trainer  # noqa: E402
 from pytorch_lightning.callbacks import ModelCheckpoint  # noqa: E402
 from tests.base import EvalModelTemplate  # noqa: E402
-from tests.base.utils import set_random_master_port, run_model_test  # noqa: E402
+from tests.base.pipelines import run_model_test  # noqa: E402
+from tests.base.utils import set_random_master_port  # noqa: E402
 
 
 parser = argparse.ArgumentParser()
